pyproject.toml
[tool.poetry]
name = "apache-spark"
version = "0.2.0"
description = "Your description goes here"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pyspark = "^3.5.1"
numpy = "^1.21.0"
pydeequ = "^1.3.0"
kafka-python = "^2.0.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"